requirements.txt
requests>=2.0.0
Pillow>=9.0.0
PyYAML>=6.0
